fix compiler warning.
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@23390 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/SVNProgressDlg.cpp
@@ -1752,7 +1752,7 @@ LRESULT CSVNProgressDlg::OnResolveMsg( WPARAM wParam, LPARAM)
     {
         for (auto it = m_arData.begin(); it != m_arData.end(); ++it)
         {
-            if ((*it)->id == wParam)
+            if ((*it)->id == (LONG)wParam)
             {
                 if ((*it)->bConflictedActionItem)
                 {
